# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "InterRPA Local API",
    "description": "Sistema offline de gestión de envíos para oficinas locales. Escaneo persistente, cola de trabajos automática y sincronización con APIs públicas.",
    "version": "0.1.0",
    "contact": {
      "name": "InterRPA Team"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3333",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/shipments/{trackingNumber}": {
      "get": {
        "summary": "Obtener guía específica",
        "description": "Obtiene los detalles de una guía por su número de rastreo",
        "operationId": "getShipmentByTracking",
        "tags": ["Shipments"],
        "parameters": [
          {
            "name": "trackingNumber",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Número de rastreo de la guía"
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles de la guía",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tracking_number": { "type": "string" },
                    "delivery_type": { "type": "string" },
                    "zone_id": { "type": ["integer", "null"] },
                    "scanned_by": { "type": "string" },
                    "office_status": { "type": "string" },
                    "recipient_name": { "type": ["string", "null"] },
                    "amount_to_collect": { "type": ["number", "null"] },
                    "payment_code": { "type": ["string", "null"] },
                    "scanned_at": { "type": "string", "format": "date-time" },
                    "updated_at": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Guía no encontrada"
          }
        }
      }
    },
    "/shipments/{trackingNumber}/jobs": {
      "get": {
        "summary": "Obtener jobs de una guía",
        "description": "Lista todos los jobs asociados a una guía específica",
        "operationId": "getShipmentJobs",
        "tags": ["Shipments"],
        "parameters": [
          {
            "name": "trackingNumber",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Número de rastreo"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de jobs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "integer" },
                      "type": { "type": "string", "enum": ["FETCH_PAYMENT_API", "FETCH_PORTAL_APX"] },
                      "tracking_number": { "type": "string" },
                      "status": { "type": "string" },
                      "attempts": { "type": "integer" },
                      "max_attempts": { "type": "integer" },
                      "run_after": { "type": "string", "format": "date-time" },
                      "last_error": { "type": ["string", "null"] },
                      "created_at": { "type": "string", "format": "date-time" },
                      "updated_at": { "type": "string", "format": "date-time" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/shipments/export": {
      "get": {
        "summary": "Exportar guías a CSV",
        "description": "Exporta guías en formato CSV con filtros opcionales por fecha, estado, zona y tipo de entrega",
        "operationId": "exportShipments",
        "tags": ["Shipments"],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": { "type": "string", "format": "date-time" },
            "description": "Fecha inicial (ISO format)"
          },
          {
            "name": "to",
            "in": "query",
            "schema": { "type": "string", "format": "date-time" },
            "description": "Fecha final (ISO format)"
          },
          {
            "name": "status",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filtrar por estado de oficina"
          },
          {
            "name": "zoneId",
            "in": "query",
            "schema": { "type": "integer" },
            "description": "Filtrar por zona"
          },
          {
            "name": "deliveryType",
            "in": "query",
            "schema": { "type": "string", "enum": ["LOCAL", "ZONA"] },
            "description": "Filtrar por tipo de entrega"
          }
        ],
        "responses": {
          "200": {
            "description": "Archivo CSV",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/jobs": {
      "get": {
        "summary": "Listar jobs",
        "description": "Obtiene jobs con filtros opcionales por estado, tipo y búsqueda de tracking",
        "operationId": "getJobs",
        "tags": ["Jobs"],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filtrar por estado (PENDING, DONE, WAITING_NET, NEEDS_HUMAN, FAILED)"
          },
          {
            "name": "type",
            "in": "query",
            "schema": { "type": "string", "enum": ["FETCH_PAYMENT_API", "FETCH_PORTAL_APX"] },
            "description": "Filtrar por tipo de job"
          },
          {
            "name": "q",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Buscar por número de rastreo (búsqueda parcial)"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 100, "maximum": 1000 },
            "description": "Límite de resultados"
          },
          {
            "name": "offset",
            "in": "query",
            "schema": { "type": "integer", "default": 0 },
            "description": "Offset para paginación"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de jobs con paginación",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": { "type": "integer" },
                        "limit": { "type": "integer" },
                        "offset": { "type": "integer" },
                        "count": { "type": "integer" }
                      }
                    },
                    "jobs": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "integer" },
                          "type": { "type": "string" },
                          "tracking_number": { "type": "string" },
                          "status": { "type": "string" },
                          "attempts": { "type": "integer" },
                          "max_attempts": { "type": "integer" },
                          "run_after": { "type": "string", "format": "date-time" },
                          "last_error": { "type": ["string", "null"] },
                          "created_at": { "type": "string", "format": "date-time" },
                          "updated_at": { "type": "string", "format": "date-time" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}